chore: simplify import for account
