+++ libs/coin-framework/src/account/balanceHistoryCache.test.ts
@@ -1,4 +1,4 @@
-import { startOfHour, startOfDay, startOfWeek } from ".";
+import { startOfHour, startOfDay, startOfWeek } from "./balanceHistoryCache";
 
 describe("date utils", () => {
   describe("Timezones", () => {
