+++ libs/coin-framework/src/account/pending.test.ts
@@ -1,8 +1,9 @@
-import { Account, Operation, TokenAccount } from "@ledgerhq/types-live";
-import { addPendingOperation } from "./pending";
 import BigNumber from "bignumber.js";
+import { Account, Operation, TokenAccount } from "@ledgerhq/types-live";
 import { listCryptoCurrencies } from "@ledgerhq/cryptoassets/index";
 import { TokenCurrency } from "@ledgerhq/types-cryptoassets";
+import { addPendingOperation } from "./pending";
+import { emptyHistoryCache } from "./balanceHistoryCache";
 
 describe("addPendingOperation", () => {
   it("add the operation to the matching account Id", () => {
@@ -62,21 +63,6 @@ describe("addPendingOperation", () => {
     expect(resultSubAcc![0].pendingOperations).not.toContain(op);
   });
 });
-
-const emptyHistoryCache = {
-  HOUR: {
-    latestDate: null,
-    balances: [],
-  },
-  DAY: {
-    latestDate: null,
-    balances: [],
-  },
-  WEEK: {
-    latestDate: null,
-    balances: [],
-  },
-};
 
 function createAccount(id: string): Account {
   const currency = listCryptoCurrencies(true)[0];
