+++ libs/coin-framework/src/mocks/account.ts
@@ -3,7 +3,8 @@ import { BigNumber } from "bignumber.js";
 import type { CryptoCurrency, TokenCurrency } from "@ledgerhq/types-cryptoassets";
 import { listCryptoCurrencies, listTokensForCryptoCurrency } from "@ledgerhq/cryptoassets/index";
 import { getOperationAmountNumber } from "../operation";
-import { isAccountEmpty, emptyHistoryCache, generateHistoryFromOperations } from "../account";
+import { isAccountEmpty, emptyHistoryCache } from "../account";
+import { generateHistoryFromOperations } from "../account/balanceHistoryCache";
 import { getDerivationScheme, runDerivationScheme } from "../derivation";
 import { genHex, genAddress } from "./helpers";
 import type { Account, AccountLike, Operation, TokenAccount } from "@ledgerhq/types-live";
