+++ libs/ledger-live-common/src/account/helpers.ts
@@ -5,56 +5,27 @@ import { TronAccount } from "../families/tron/types";
 import { CosmosAccount } from "../families/cosmos/types";
 import {
   getMainAccount,
-  getAccountCurrency,
-  getFeesCurrency,
-  getFeesUnit,
-  getAccountSpendableBalance,
   isAccountEmpty as commonIsAccountEmpty,
-  areAllOperationsLoaded,
-  isAccountBalanceSignificant,
   clearAccount as commonClearAccount,
-  findSubAccountById,
-  listSubAccounts,
-  FlattenAccountsOptions,
-  flattenAccounts,
-  shortAddressPreview,
-  isAccountBalanceUnconfirmed,
-  isUpToDateAccount,
-  makeEmptyTokenAccount,
-  accountWithMandatoryTokens,
-  withoutToken,
-  findTokenAccountByCurrency,
-  isAccount,
-  isTokenAccount,
-  isSubAccount,
-  getParentAccount,
-} from "@ledgerhq/coin-framework/account/helpers";
+} from "@ledgerhq/coin-framework/account/index";
 
-// FIXME Remove this redundant export and import all from coin-framework
+// TODO: remove this export and prefer import from root file.
 export {
+  accountWithMandatoryTokens,
+  findTokenAccountByCurrency,
+  flattenAccounts,
   getMainAccount,
   getAccountCurrency,
+  getAccountSpendableBalance,
   getFeesCurrency,
   getFeesUnit,
-  getAccountSpendableBalance,
-  areAllOperationsLoaded,
-  isAccountBalanceSignificant,
-  findSubAccountById,
-  listSubAccounts,
-  FlattenAccountsOptions,
-  flattenAccounts,
-  shortAddressPreview,
-  isAccountBalanceUnconfirmed,
-  isUpToDateAccount,
-  makeEmptyTokenAccount,
-  accountWithMandatoryTokens,
-  withoutToken,
-  findTokenAccountByCurrency,
-  isAccount,
-  isTokenAccount,
-  isSubAccount,
   getParentAccount,
-};
+  isAccount,
+  isAccountBalanceUnconfirmed,
+  isTokenAccount,
+  listSubAccounts,
+  shortAddressPreview,
+} from "@ledgerhq/coin-framework/account/index";
 
 export const isAccountEmpty = (a: AccountLike): boolean => {
   return commonIsAccountEmpty(a);
@@ -112,3 +83,15 @@ export const getVotesCount = (
       return 0;
   }
 };
+
+export function areAllOperationsLoaded(account: AccountLike): boolean {
+  if (account.operationsCount !== account.operations.length) {
+    return false;
+  }
+
+  if (account.type === "Account" && account.subAccounts) {
+    return account.subAccounts.every(areAllOperationsLoaded);
+  }
+
+  return true;
+}
