+++ libs/ledger-live-common/src/account/index.ts
@@ -1,8 +1,17 @@
-export * from "./helpers";
-export * from "./support";
-export * from "./serialization";
-export * from "@ledgerhq/coin-framework/account/accountId";
-export * from "@ledgerhq/coin-framework/account/groupOperations";
-export * from "@ledgerhq/coin-framework/account/pending";
+export {
+  encodeAccountId,
+  decodeAccountId,
+  encodeTokenAccountId,
+  decodeTokenAccountId,
+  isUpToDateAccount,
+  makeEmptyTokenAccount,
+  findSubAccountById,
+  type FlattenAccountsOptions,
+  groupAccountOperationsByDay,
+  groupAccountsOperationsByDay,
+  addPendingOperation,
+} from "@ledgerhq/coin-framework/account/index";
 export * from "./formatters";
-export * from "@ledgerhq/coin-framework/account/balanceHistoryCache";
+export * from "./helpers";
+export * from "./serialization";
+export * from "./support";
