+++ libs/ledger-live-common/src/account/support.ts
@@ -2,13 +2,10 @@ import { getMainAccount } from "@ledgerhq/coin-framework/account/index";
 import { getAccountBridge } from "../bridge";
 import type { Account, AccountLike } from "@ledgerhq/types-live";
 
-import {
-  shouldShowNewAccount,
+export {
   getReceiveFlowError,
   checkAccountSupported,
 } from "@ledgerhq/coin-framework/account/support";
-
-export { shouldShowNewAccount, getReceiveFlowError, checkAccountSupported };
 
 export function canSend(account: AccountLike, parentAccount: Account | null | undefined): boolean {
   try {
