+++ libs/ledger-live-common/src/exchange/swap/getUpdateAccountWithUpdaterParams.ts
@@ -2,8 +2,7 @@ import { BigNumber } from "bignumber.js";
 import { Account, Operation } from "@ledgerhq/types-live";
 import addToSwapHistory from "./addToSwapHistory";
 import { ExchangeSwap, ExchangeRate } from "./types";
-import { addPendingOperation } from "../../account/index";
-import { getMainAccount } from "../../account/helpers";
+import { addPendingOperation, getMainAccount } from "../../account/index";
 import { Transaction } from "../../generated/types";
 
 export const getUpdateAccountWithUpdaterParams = ({
