+++ libs/ledger-live-common/src/families/elrond/synchronisation.integration.test.ts
@@ -1,7 +1,6 @@
 import "../../__tests__/test-helpers/setup";
 import { reduce } from "rxjs/operators";
-import { fromAccountRaw } from "../../account";
-import { getAccountCurrency } from "../../account/helpers";
+import { fromAccountRaw, getAccountCurrency } from "../../account";
 import type { Account, SubAccount } from "@ledgerhq/types-live";
 import { getAccountBridge } from "../../bridge";
 import { makeBridgeCacheSystem } from "../../bridge/cache";
